create a cacheable version of predict_rc
